requirements.txt
flask==2.0.1
weaviate-client==3.11.0
python-dotenv==0.19.0
PyPDF2==3.0.1
reportlab==3.6.8
beautifulsoup4==4.9.3
requests==2.26.0
pymongo==4.5.0
Werkzeug==2.0.1
crewai==0.11.2
python-docx==0.8.11